+++ src/components/content/VideoPage/VideoPlayer/VideoPlayerDescription.js
@@ -1,30 +1,7 @@
 import React, { Component } from 'react';
+import formatDate from '../../../../helper';
 
 class VideoPlayerDescription extends Component {
-  static formatDate(publishedAt) {
-    const dateObj = new Date(publishedAt);
-
-    const monthNames = [
-      'January',
-      'February',
-      'March',
-      'April',
-      'May',
-      'June',
-      'July',
-      'August',
-      'September',
-      'October',
-      'November',
-      'December',
-    ];
-
-    const month = monthNames[dateObj.getMonth()];
-    const day = dateObj.getUTCDate();
-    const year = dateObj.getUTCFullYear();
-
-    return `Published on ${month} ${day}, ${year}`;
-  }
 
   render() {
     const { channelTitle, description, publishedAt } = this.props;
@@ -36,7 +13,7 @@ class VideoPlayerDescription extends Component {
         </div>
         <div className="description">
           <h2>{channelTitle}</h2>
-          <h3>{VideoPlayerDescription.formatDate(publishedAt)}</h3>
+          <h3>{formatDate(publishedAt)}</h3>
           <p>{description}</p>
           <p className="show-more">show more</p>
         </div>
